requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
openai==1.76.0
pydantic==2.11.3
pymongo==4.12.0
pyngrok==7.2.3
python-dotenv==1.1.0
Requests==2.32.3
uvicorn==0.34.2
